+++ changelog.md
@@ -9,6 +9,7 @@
 - ADD: Configure Compile Command under <SETTINGS>
 - ADD: Restart button under <FILE> Menu
 - ADD: Button to stop the current compile.
+- DEL: Does not show Button hint in Statusbar.
 
 ## v0.1.0